# performance improvement - isTrue, isFalseTest, SourceLineInfo ctor