@horenmar 2017-11-20
Are you also checking what happens with compilation times?