@coombez 2017-11-20
I do not have a good benchmark for the compile times as I do for run time, but I am paying attention to make sure the compile time doesn't explode.  Trivial changes like these wouldn't typically affect the compile time, but measuring is always better than intuition :)

Have you guys thought about adding benchmarks for the stuff you care about to your test suite?