@coombez 2017-11-20
ha!  Just noticed the benchmarkCompile.py.  I'll check it out and add it to my workflow