@coombez 2017-11-20
Hmm.  I'm not sure how to use benchmarkCompile.py out of the box...  I'll dig deeper when I have time.
OSError: [Errno 2] No such file or directory: 'benchmark-dir'