@philsquared 2017-11-20
This is a tricky one. IIRC `isTrue()` was non-inlined so the compiler wouldn't complain about unreachable code. However that was always a bit of an arms race and may be best left to `#pragma`s - in which case it may not be needed at all in most places. Inlining it is solving the wrong problem, though, I think