@horenmar 2017-11-20
@coombez Yeah, `benchmarkCompily.py` is a bit idiosyncratic, I mostly created that one for myself.

Anyway, rough usage guide:
* You need to generate the test files first. This means running it with `-g` or `--generate-files` and `-I <path-to-single-include>`.
* You should now see `benchmark-dir` folder where you ran the script, it should contain `catch.hpp` and `.cpp` files to test compilation
* Run `benchmarkCompile.py` again, this time with an argument specifying which type of compilation you want to check (obviously link requires files to be compiled first). You will also likely need to specify `-f=-std=c++11`.