@coombez 2017-11-20
@philsquared All the existing checks pass... does this mean the issue with isTrue() is not covered? or that isTrue() is not needed? or something else all together?  Obviously I don't fundamentally care about it being inlined, it was just consuming 10% of the runtime so it seems like a simple win.

Just as an added data point, in Catch 1.9.6, isTrue is in fact inlined with the same compiler warning comment that it still has.  Perhaps this is still good enough?