@martinmoene 2017-11-20
Note: where Catch uses
```
while( Catch::isTrue( false && static_cast<bool>( !!(__VA_ARGS__) ) ) )
```
doctest uses 
```
while((void)0, 0)
```