@horenmar 2017-11-21
@martinmoene That formulation suppresses warnings from tested code. Consider these three files

`a.cpp`
```cpp
#include <iostream>

int main() {
    int i = 1;
    if (i == NULL) {
        std::cout << "i = " << i << std::endl;
    }
}
```
`b.cpp`
```cpp
#define CATCH_CONFIG_MAIN
#include "catch.hpp"

TEST_CASE("foo") {
    int i = 1;
    REQUIRE(i == NULL);
}
```
`c.cpp`
```cpp
#define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
#include "doctest.h"

TEST_CASE("foo") {
    int i = 1;
    CHECK(i == NULL);
}
```

when compiled with `-Wall -Wextra`, both `a.cpp` and `b.cpp` warn about comparing `int` with `NULL`. `c.cpp` does not.