@horenmar 2017-11-21
We could probably throw out the `IsTrue`, but g++'s handling of `_Pragma`s is buggy as hell, so it might mean having to suppress tautological compare in the whole TU instead.