@martinmoene 2017-11-21
@horenmar a squint owl that still keeps a sharp look:
```
while( (void)0, 0 && static_cast<bool>( !!(__VA_ARGS__) ) )
```