@philsquared 2017-11-21
I get warnings with `(void)0`. However I tested a re-inlined `isTrue()` on a few GCCs, clang and MSVC (2017) with high(ish) warnings (-wall/ -W4) and get no warnings from it - so I've now committed that.