@philsquared 2017-11-21
will look at the others when I have a bit more time to run the compiler benchmarks as there can be an interplay.