@martinmoene 2017-11-21
Did you really try `while( (void)0,0 && ...` (note the `0,0` ) ?

I've not seen (undesired) warnings with vs2015/2017 -W4, or various gcc/clang -Wall.