@philsquared 2017-11-21
@martinmoene hmmm, odd. I just tried it again and this time get no warning with `,` (but I do with `&&`).
I wonder if I left the `&&` in there after previously trying both ways with the other compilers (all in Compiler Explorer).
Maybe that *is* better, then...