@coombez 2017-11-21
Results from benchmarkCompile.py:

Before:
Time needed for ...
        ... compiling main, mean: 4.34, median: 4.36 s
        ... compiling test files, mean: 18.68, median: 18.66 s
        ... linking everything, mean: 0.22, median: 0.21 s

After:
Time needed for ...
        ... compiling main, mean: 5.78, median: 5.75 s
        ... compiling test files, mean: 18.63, median: 18.62 s
        ... linking everything, mean: 0.28, median: 0.28 s