@coombez 2017-11-21
I'll have to re-run the benchmarks, that was from the checked in generated header to the generated header after these changes.  I should run the benchmark against the generated header just before these changes instead.