@horenmar 2017-12-09
I took the liberty of manually picking parts of this that are still relevant. This means no credit for you (sorry), but it seemed like the least painful option.